+++ src/app/reports/reporting-status-reports/page.tsx
@@ -6,10 +6,8 @@ import { message } from "antd";
 import outputs from "@root/amplify_outputs.json";
 import Page from "@/shared/ui/page/ui/Page";
 import { ReportingStatusReportsList } from "@/widgets/reporting-status-reports";
-import { ExportProjectReportButton } from "@/feature/export-project-reports";
 import { useState } from "react";
 import { ReportingStatusReport } from "@/entities/reporting-status-reports/config/types";
-import { data } from "@/entities/reporting-status-reports/api/reporting-status-reports";
 import { ExportMonthlyProjectReportButton } from "@/feature/export-project-monthly-reports";
 Amplify.configure(outputs);
