+++ src/entities/reporting-status-reports/api/reporting-status-reports-jwdnjn.ts
@@ -1,0 +1,692 @@
+export const data = [
+  {
+    id: 1,
+    project: "Project Alpha",
+    cluster: "Kerala",
+    region: "South",
+    facilitator: "Facilitator 1",
+    status: "Approved",
+    date: "23/12/2025",
+    reportingMonth: "December",
+    year: "2024",
+    goalsFromLastMonth: [
+      {
+        goal: "Increase local engagement",
+        majorGoal: true,
+        functionalArea: "Community Outreach",
+        achieved: "Yes",
+        reason: "Strong volunteer participation",
+        comments: "Exceeded expectations",
+      },
+      {
+        goal: "Host a workshop",
+        majorGoal: false,
+        functionalArea: "Education",
+        achieved: "No",
+        reason: "Scheduling conflicts",
+        comments: "Will reschedule next month",
+      },
+      {
+        goal: "Develop new partnerships",
+        majorGoal: true,
+        functionalArea: "Partnerships",
+        achieved: "Yes",
+        reason: "Effective networking",
+        comments: "Two new MOUs signed",
+      },
+    ],
+    additionalActivities: [
+      {
+        goal: "Organize a fundraising event",
+        majorGoal: true,
+        functionalArea: "Fundraising",
+        achieved: "Yes",
+        reason: "High donor turnout",
+        comments: "Raised 20% more than target",
+      },
+      {
+        goal: "Launch awareness campaign",
+        majorGoal: true,
+        functionalArea: "Marketing",
+        achieved: "No",
+        reason: "Budget constraints",
+        comments: "Postponed to next quarter",
+      },
+      {
+        goal: "Recruit volunteers",
+        majorGoal: false,
+        functionalArea: "HR",
+        achieved: "Yes",
+        reason: "Successful outreach",
+        comments: "Added 15 new volunteers",
+      },
+    ],
+    nextMonthGoal: [
+      {
+        goal: "Conduct training sessions",
+        majorGoal: true,
+        functionalArea: "Education",
+        comments: "Focus on digital skills",
+      },
+      {
+        goal: "Improve donor retention",
+        majorGoal: false,
+        functionalArea: "Fundraising",
+        comments: "Send personalized thank-you notes",
+      },
+      {
+        goal: "Expand social media reach",
+        majorGoal: false,
+        functionalArea: "Marketing",
+        comments: "Target younger demographics",
+      },
+    ],
+    nextMonthAdditional: [
+      {
+        goal: "Organize health camp",
+        majorGoal: false,
+        functionalArea: "Community Health",
+        comments: "Collaborate with local hospitals",
+      },
+      {
+        goal: "Introduce feedback mechanisms",
+        majorGoal: false,
+        functionalArea: "Quality Assurance",
+        comments: "Deploy online surveys",
+      },
+      {
+        goal: "Host community cleanup drive",
+        majorGoal: false,
+        functionalArea: "Environment",
+        comments: "Engage local schools",
+      },
+    ],
+    prayerRequests: [
+      "Pray for continued volunteer engagement",
+      "Pray for smooth rescheduling of the workshop",
+    ],
+    praisePoints: [
+      "Praise for high donor turnout in the fundraising event",
+      "Praise for the new partnerships formed",
+    ],
+    story:
+      "A testimony of how our outreach efforts led to two significant partnerships that will benefit the community.",
+    concerns:
+      "Concerns about the postponed awareness campaign and the budget constraints.",
+  },
+  {
+    id: 2,
+    project: "Project Beta",
+    cluster: "Punjab",
+    region: "North",
+    facilitator: "Facilitator 2",
+    status: "In Progress",
+    date: "20/11/2025",
+    reportingMonth: "November",
+    year: "2024",
+    goalsFromLastMonth: [
+      {
+        goal: "Launch new website",
+        majorGoal: true,
+        functionalArea: "IT",
+        achieved: "Yes",
+        reason: "Met deadlines",
+        comments: "Website live and functional",
+      },
+      {
+        goal: "Provide basic education kits",
+        majorGoal: true,
+        functionalArea: "Education",
+        achieved: "No",
+        reason: "Logistical delays",
+        comments: "Awaiting shipment",
+      },
+      {
+        goal: "Train new staff",
+        majorGoal: true,
+        functionalArea: "HR",
+        achieved: "Yes",
+        reason: "Training sessions completed",
+        comments: "Positive feedback received",
+      },
+    ],
+    additionalActivities: [
+      {
+        goal: "Conduct baseline survey",
+        majorGoal: true,
+        functionalArea: "Research",
+        achieved: "Yes",
+        reason: "Team efforts",
+        comments: "Data collection complete",
+      },
+      {
+        goal: "Enhance data security",
+        majorGoal: true,
+        functionalArea: "IT",
+        achieved: "No",
+        reason: "Resource constraints",
+        comments: "Will prioritize next quarter",
+      },
+      {
+        goal: "Collaborate with NGOs",
+        majorGoal: true,
+        functionalArea: "Partnerships",
+        achieved: "Yes",
+        reason: "Shared objectives",
+        comments: "Five partnerships established",
+      },
+    ],
+    nextMonthGoal: [
+      {
+        goal: "Initiate teacher training",
+        majorGoal: true,
+        functionalArea: "Education",
+        comments: "Focus on STEM subjects",
+      },
+      {
+        goal: "Promote environmental awareness",
+        majorGoal: true,
+        functionalArea: "Environment",
+        comments: "Run workshops in schools",
+      },
+      {
+        goal: "Host inter-department meeting",
+        majorGoal: true,
+        functionalArea: "Management",
+        comments: "Align goals across teams",
+      },
+    ],
+    nextMonthAdditional: [
+      {
+        goal: "Plan annual fundraiser",
+        majorGoal: true,
+        functionalArea: "Fundraising",
+        comments: "Secure venue and sponsors",
+      },
+      {
+        goal: "Upgrade IT infrastructure",
+        majorGoal: true,
+        functionalArea: "IT",
+        comments: "Replace outdated hardware",
+      },
+      {
+        goal: "Evaluate program impact",
+        majorGoal: true,
+        functionalArea: "Quality Assurance",
+        comments: "Analyze feedback data",
+      },
+    ],
+    prayerRequests: [
+      "Pray for the logistical issues surrounding the education kits",
+      "Pray for resources to enhance data security",
+    ],
+    praisePoints: [
+      "Praise for the successful website launch",
+      "Praise for the new staff training program",
+    ],
+    story:
+      "Our successful launch of the website provided a new platform for outreach and engagement, opening doors for more interaction.",
+    concerns:
+      "Concerns regarding resource constraints for the IT infrastructure upgrade and delays in education kit shipment.",
+  },
+  {
+    id: 3,
+    project: "Project Gamma",
+    cluster: "Tamil Nadu",
+    region: "South",
+    facilitator: "Facilitator 3",
+    status: "Completed",
+    date: "15/10/2025",
+    reportingMonth: "October",
+    year: "2024",
+    goalsFromLastMonth: [
+      {
+        goal: "Setup local office",
+        majorGoal: true,
+        functionalArea: "Operations",
+        achieved: "Yes",
+        reason: "Timely planning",
+        comments: "Office operational",
+      },
+      {
+        goal: "Distribute study material",
+        majorGoal: true,
+        functionalArea: "Education",
+        achieved: "Yes",
+        reason: "Good logistics",
+        comments: "Completed within budget",
+      },
+      {
+        goal: "Train volunteers",
+        majorGoal: true,
+        functionalArea: "HR",
+        achieved: "No",
+        reason: "Low attendance",
+        comments: "Plan to reattempt",
+      },
+    ],
+    additionalActivities: [
+      {
+        goal: "Implement new CRM",
+        majorGoal: true,
+        functionalArea: "IT",
+        achieved: "Yes",
+        reason: "Smooth rollout",
+        comments: "CRM in use",
+      },
+      {
+        goal: "Organize sports day",
+        majorGoal: true,
+        functionalArea: "Community",
+        achieved: "Yes",
+        reason: "High turnout",
+        comments: "Event a success",
+      },
+      {
+        goal: "Improve accessibility",
+        majorGoal: true,
+        functionalArea: "Infrastructure",
+        achieved: "No",
+        reason: "Budget issues",
+        comments: "Reschedule for next quarter",
+      },
+    ],
+    nextMonthGoal: [
+      {
+        goal: "Launch mobile app",
+        majorGoal: true,
+        functionalArea: "IT",
+        comments: "Targeting better user experience",
+      },
+      {
+        goal: "Enhance curriculum",
+        majorGoal: true,
+        functionalArea: "Education",
+        comments: "Include interactive modules",
+      },
+      {
+        goal: "Host recruitment drive",
+        majorGoal: true,
+        functionalArea: "HR",
+        comments: "Focus on key roles",
+      },
+    ],
+    nextMonthAdditional: [
+      {
+        goal: "Plan cultural event",
+        majorGoal: true,
+        functionalArea: "Community",
+        comments: "Promote local traditions",
+      },
+      {
+        goal: "Upgrade training materials",
+        majorGoal: true,
+        functionalArea: "Education",
+        comments: "Include online resources",
+      },
+      {
+        goal: "Set up grievance cell",
+        majorGoal: true,
+        functionalArea: "Operations",
+        comments: "Ensure better feedback",
+      },
+    ],
+    prayerRequests: [
+      "Pray for the volunteer training reattempt",
+      "Pray for budget to improve accessibility in the next quarter",
+    ],
+    praisePoints: [
+      "Praise for the successful implementation of the CRM system",
+      "Praise for the sports day event success",
+    ],
+    story:
+      "The sports day was a huge success, drawing many community members together and fostering a sense of belonging.",
+    concerns:
+      "Concerns about the low attendance during volunteer training and accessibility improvements delayed by budget constraints.",
+  },
+  {
+    id: 4,
+    project: "Project Alpha",
+    cluster: "Maharashtra",
+    region: "West",
+    facilitator: "Facilitator 4",
+    status: "In Progress",
+    date: "10/01/2025",
+    reportingMonth: "January",
+    year: "2025",
+    goalsFromLastMonth: [
+      {
+        goal: "Increase volunteer participation",
+        majorGoal: true,
+        functionalArea: "Community Outreach",
+        achieved: "Yes",
+        reason: "Effective engagement strategy",
+        comments: "Successful outreach program",
+      },
+      {
+        goal: "Set up awareness campaign",
+        majorGoal: true,
+        functionalArea: "Marketing",
+        achieved: "No",
+        reason: "Delay in materials",
+        comments: "Materials arriving next week",
+      },
+      {
+        goal: "Establish new partnerships",
+        majorGoal: true,
+        functionalArea: "Partnerships",
+        achieved: "Yes",
+        reason: "Strong networking efforts",
+        comments: "Four new partnerships formed",
+      },
+    ],
+    additionalActivities: [
+      {
+        goal: "Host webinar on community health",
+        majorGoal: true,
+        functionalArea: "Community Health",
+        achieved: "Yes",
+        reason: "High attendance",
+        comments: "Positive feedback from participants",
+      },
+      {
+        goal: "Launch digital donation platform",
+        majorGoal: true,
+        functionalArea: "Fundraising",
+        achieved: "No",
+        reason: "Technical delays",
+        comments: "Platform expected to launch next month",
+      },
+      {
+        goal: "Provide skill-building workshops",
+        majorGoal: true,
+        functionalArea: "Education",
+        achieved: "Yes",
+        reason: "Excellent turnout",
+        comments: "15 workshops conducted",
+      },
+    ],
+    nextMonthGoal: [
+      {
+        goal: "Expand volunteer program",
+        majorGoal: true,
+        functionalArea: "HR",
+        comments: "Targeting youth engagement",
+      },
+      {
+        goal: "Enhance social media strategy",
+        majorGoal: true,
+        functionalArea: "Marketing",
+        comments: "Focus on Instagram and TikTok",
+      },
+      {
+        goal: "Improve donor engagement",
+        majorGoal: true,
+        functionalArea: "Fundraising",
+        comments: "Host a virtual thank-you event",
+      },
+    ],
+    nextMonthAdditional: [
+      {
+        goal: "Initiate youth leadership program",
+        majorGoal: true,
+        functionalArea: "Community Outreach",
+        comments: "Develop training materials",
+      },
+      {
+        goal: "Evaluate environmental impact",
+        majorGoal: true,
+        functionalArea: "Environment",
+        comments: "Conduct baseline assessment",
+      },
+      {
+        goal: "Upgrade website functionality",
+        majorGoal: true,
+        functionalArea: "IT",
+        comments: "Improve user experience",
+      },
+    ],
+    prayerRequests: [
+      "Pray for the successful expansion of the volunteer program",
+      "Pray for the digital donation platform launch to go smoothly",
+    ],
+    praisePoints: [
+      "Praise for the positive feedback from the community health webinar",
+      "Praise for the four new partnerships formed",
+    ],
+    story:
+      "We received overwhelming support from the community during the health webinar, which opened the door for new collaboration opportunities.",
+    concerns:
+      "Concerns over the delay in materials for the awareness campaign and the technical issues delaying the donation platform launch.",
+  },
+  {
+    id: 5,
+    project: "Project Beta",
+    cluster: "Uttar Pradesh",
+    region: "North",
+    facilitator: "Facilitator 5",
+    status: "Approved",
+    date: "25/11/2025",
+    reportingMonth: "November",
+    year: "2024",
+    goalsFromLastMonth: [
+      {
+        goal: "Launch community health program",
+        majorGoal: true,
+        functionalArea: "Community Health",
+        achieved: "Yes",
+        reason: "Strong partnership with local hospitals",
+        comments: "Successfully launched with 200 attendees",
+      },
+      {
+        goal: "Provide educational scholarships",
+        majorGoal: true,
+        functionalArea: "Education",
+        achieved: "No",
+        reason: "Delay in approval process",
+        comments: "Scholarships will be disbursed next month",
+      },
+      {
+        goal: "Set up mental health support group",
+        majorGoal: true,
+        functionalArea: "Health",
+        achieved: "Yes",
+        reason: "Well-received by the community",
+        comments: "Group sessions ongoing",
+      },
+    ],
+    additionalActivities: [
+      {
+        goal: "Establish school partnership",
+        majorGoal: true,
+        functionalArea: "Education",
+        achieved: "Yes",
+        reason: "Agreement finalized",
+        comments: "Five schools onboarded",
+      },
+      {
+        goal: "Launch mobile health app",
+        majorGoal: true,
+        functionalArea: "IT",
+        achieved: "No",
+        reason: "App under review",
+        comments: "Launch scheduled for next quarter",
+      },
+      {
+        goal: "Organize local talent show",
+        majorGoal: false,
+        functionalArea: "Community Engagement",
+        achieved: "Yes",
+        reason: "Huge success",
+        comments: "300 attendees and 20 performers",
+      },
+    ],
+    nextMonthGoal: [
+      {
+        goal: "Expand scholarship program",
+        majorGoal: false,
+        functionalArea: "Education",
+        comments: "Targeting students from remote areas",
+      },
+      {
+        goal: "Launch mental health awareness campaign",
+        majorGoal: false,
+        functionalArea: "Health",
+        comments: "Focus on schools and colleges",
+      },
+      {
+        goal: "Enhance community engagement",
+        majorGoal: false,
+        functionalArea: "Community Outreach",
+        comments: "Host neighborhood meetups",
+      },
+    ],
+    nextMonthAdditional: [
+      {
+        goal: "Introduce wellness program",
+        majorGoal: false,
+        functionalArea: "Community Health",
+        comments: "Partner with fitness trainers",
+      },
+      {
+        goal: "Host leadership workshop",
+        majorGoal: false,
+        functionalArea: "HR",
+        comments: "Develop training modules",
+      },
+      {
+        goal: "Improve event logistics",
+        majorGoal: false,
+        functionalArea: "Operations",
+        comments: "Ensure smooth execution",
+      },
+    ],
+    prayerRequests: [
+      "Pray for the scholarship program to reach more students",
+      "Pray for the mobile health app review to go smoothly",
+    ],
+    praisePoints: [
+      "Praise for the successful local talent show",
+      "Praise for the new school partnerships",
+    ],
+    story:
+      "The local talent show was a big hit, bringing the community together and showcasing some incredible talent.",
+    concerns:
+      "Concerns about the delay in the mobile health app review and challenges in expanding the scholarship program.",
+  },
+  {
+    id: 6,
+    project: "Project Gamma",
+    cluster: "Rajasthan",
+    region: "West",
+    facilitator: "Facilitator 6",
+    status: "Completed",
+    date: "30/09/2025",
+    reportingMonth: "September",
+    year: "2024",
+    goalsFromLastMonth: [
+      {
+        goal: "Launch community garden project",
+        majorGoal: false,
+        functionalArea: "Environment",
+        achieved: "Yes",
+        reason: "Community involvement",
+        comments: "Successfully planted 500 trees",
+      },
+      {
+        goal: "Provide access to clean water",
+        majorGoal: false,
+        functionalArea: "Community Health",
+        achieved: "No",
+        reason: "Delayed equipment shipment",
+        comments: "Expected next month",
+      },
+      {
+        goal: "Organize leadership seminar",
+        majorGoal: false,
+        functionalArea: "Education",
+        achieved: "Yes",
+        reason: "High demand",
+        comments: "Seminar attended by 150 people",
+      },
+    ],
+    additionalActivities: [
+      {
+        goal: "Promote sustainable farming practices",
+        majorGoal: false,
+        functionalArea: "Agriculture",
+        achieved: "Yes",
+        reason: "Positive community response",
+        comments: "20 farmers trained",
+      },
+      {
+        goal: "Develop waste management plan",
+        majorGoal: false,
+        functionalArea: "Environment",
+        achieved: "No",
+        reason: "Budget constraints",
+        comments: "Plan delayed until next quarter",
+      },
+      {
+        goal: "Create youth mentorship program",
+        majorGoal: false,
+        functionalArea: "HR",
+        achieved: "Yes",
+        reason: "Strong mentor interest",
+        comments: "15 mentors enrolled",
+      },
+    ],
+    nextMonthGoal: [
+      {
+        goal: "Launch waste segregation initiative",
+        majorGoal: false,
+        functionalArea: "Environment",
+        comments: "Collaborate with local authorities",
+      },
+      {
+        goal: "Host awareness sessions on clean water",
+        majorGoal: false,
+        functionalArea: "Community Health",
+        comments: "Partner with NGOs",
+      },
+      {
+        goal: "Expand youth mentorship program",
+        majorGoal: false,
+        functionalArea: "HR",
+        comments: "Increase mentor recruitment",
+      },
+    ],
+    nextMonthAdditional: [
+      {
+        goal: "Implement solar energy solutions",
+        majorGoal: false,
+        functionalArea: "Environment",
+        comments: "Focus on rural areas",
+      },
+      {
+        goal: "Provide vocational training",
+        majorGoal: false,
+        functionalArea: "Education",
+        comments: "Develop curriculum for trades",
+      },
+      {
+        goal: "Facilitate leadership training",
+        majorGoal: false,
+        functionalArea: "Community Outreach",
+        comments: "Develop interactive modules",
+      },
+    ],
+    prayerRequests: [
+      "Pray for the timely arrival of clean water equipment",
+      "Pray for the successful launch of the waste segregation initiative",
+    ],
+    praisePoints: [
+      "Praise for the community garden project",
+      "Praise for the leadership seminar success",
+    ],
+    story:
+      "The community garden has transformed the local landscape and is a beautiful example of sustainable practices in action.",
+    concerns:
+      "Concerns over the delay in the waste management plan and the clean water project equipment delay.",
+  },
+];
