+++ src/feature/export-miscellaneous-reports/ui/ExportMiscellaneousReportButton.tsx
@@ -18,14 +18,19 @@ export default function ExportMiscellaneousReportButton({
   month,
 }: ExportMiscellaneousReportButtonProps) {
   const exportDocument = (projectReports: MiscellaneousProjectReport[]) => {
-    console.log("Misc title is ", miscTitle);
     const groupedReports = projectReports.reduce((acc, report) => {
-      if (!acc[report.project]) {
-        acc[report.project] = [];
+      const projectName = report.project;
+      const facilitatorName = report.facilitator;
+
+      if (!acc[projectName]) {
+        acc[projectName] = {};
       }
-      acc[report.project].push(report);
+      if (!acc[projectName][facilitatorName]) {
+        acc[projectName][facilitatorName] = [];
+      }
+      acc[projectName][facilitatorName].push(report);
       return acc;
-    }, {} as Record<string, MiscellaneousProjectReport[]>);
+    }, {} as Record<string, Record<string, MiscellaneousProjectReport[]>>);
 
     const documentContent = [
       new Paragraph({
@@ -40,7 +45,6 @@ export default function ExportMiscellaneousReportButton({
           }),
         ],
       }),
-
       new Paragraph({
         text: `Month: ${month}`,
         spacing: { after: 200 },
@@ -52,155 +56,177 @@ export default function ExportMiscellaneousReportButton({
         spacing: { after: 200 },
       }),
 
-      // Loop through projects and create content for each project
-      ...Object.entries(groupedReports).flatMap(([projectName, reports]) => [
-        new Paragraph({
-          children: [new TextRun({ text: projectName, size: 24, bold: true })],
-          alignment: AlignmentType.CENTER,
-          spacing: { after: 200 },
-        }),
-
-        new Paragraph({
-          text: "───────────────────────────────────────────────────────────────",
-          alignment: AlignmentType.CENTER,
-          spacing: { after: 100 },
-        }),
-
-        // Conditional sections based on miscTitle
-        ...reports.flatMap((report) => {
-          const sections: Paragraph[] = [];
-
-          if (miscTitle === "Praise points/Prayer request") {
-            sections.push(
+      ...Object.entries(groupedReports).flatMap(
+        ([projectName, facilitators]) => [
+          new Paragraph({
+            children: [
+              new TextRun({ text: projectName, size: 24, bold: true }),
+            ],
+            alignment: AlignmentType.CENTER,
+            spacing: { after: 200 },
+          }),
+          new Paragraph({
+            text: "───────────────────────────────────────────────────────────────",
+            alignment: AlignmentType.CENTER,
+            spacing: { after: 100 },
+          }),
+          ...Object.entries(facilitators).flatMap(
+            ([facilitatorName, reports]) => [
               new Paragraph({
                 children: [
                   new TextRun({
-                    text: "Praise points / Prayer requests",
-                    size: 18,
+                    text: `Facilitator: ${facilitatorName}`,
+                    size: 20,
                     bold: true,
                   }),
                 ],
-                heading: "Heading2",
-                spacing: { after: 200 },
+                spacing: { after: 150 },
               }),
-              new Paragraph({
-                children: [
-                  new TextRun({
-                    text: "Praise for",
-                    size: 16,
-                    bold: true,
-                  }),
-                ],
-                heading: "Heading3",
-                spacing: { after: 200 },
-              }),
-              ...(report.praisePoints?.length
-                ? report.praisePoints.map(
-                    (point) =>
-                      new Paragraph({
-                        children: [new TextRun({ text: point, size: 14 })],
-                        bullet: { level: 1 },
-                        spacing: { after: 200 },
-                      })
-                  )
-                : [
+
+              ...reports.flatMap((report) => {
+                const sections: Paragraph[] = [];
+
+                if (miscTitle === "Praise points/Prayer request") {
+                  sections.push(
                     new Paragraph({
                       children: [
                         new TextRun({
-                          text: "No praise points available.",
+                          text: "Praise points / Prayer requests",
+                          size: 18,
+                          bold: true,
+                        }),
+                      ],
+                      heading: "Heading2",
+                      spacing: { after: 200 },
+                    }),
+                    new Paragraph({
+                      children: [
+                        new TextRun({
+                          text: "Praise for",
+                          size: 16,
+                          bold: true,
+                        }),
+                      ],
+                      heading: "Heading3",
+                      spacing: { after: 200 },
+                    }),
+                    ...(report.praisePoints?.length
+                      ? report.praisePoints.map(
+                          (point) =>
+                            new Paragraph({
+                              children: [
+                                new TextRun({ text: point, size: 14 }),
+                              ],
+                              bullet: { level: 1 },
+                              spacing: { after: 200 },
+                            })
+                        )
+                      : [
+                          new Paragraph({
+                            children: [
+                              new TextRun({
+                                text: "No praise points available.",
+                                size: 14,
+                              }),
+                            ],
+                            spacing: { after: 200 },
+                          }),
+                        ]),
+                    new Paragraph({
+                      children: [
+                        new TextRun({
+                          text: "Pray for",
+                          size: 16,
+                          bold: true,
+                        }),
+                      ],
+                      heading: "Heading3",
+                      spacing: { after: 200 },
+                    }),
+                    ...(report.prayerRequests?.length
+                      ? report.prayerRequests.map(
+                          (request) =>
+                            new Paragraph({
+                              children: [
+                                new TextRun({ text: request, size: 14 }),
+                              ],
+                              bullet: { level: 1 },
+                              spacing: { after: 200 },
+                            })
+                        )
+                      : [
+                          new Paragraph({
+                            children: [
+                              new TextRun({
+                                text: "No prayer requests available.",
+                                size: 14,
+                              }),
+                            ],
+                            spacing: { after: 200 },
+                          }),
+                        ])
+                  );
+                }
+
+                if (miscTitle === "Concerns/Struggles") {
+                  sections.push(
+                    new Paragraph({
+                      children: [
+                        new TextRun({
+                          text: "Concerns / Struggles",
+                          size: 18,
+                          bold: true,
+                        }),
+                      ],
+                      heading: "Heading2",
+                      spacing: { after: 200 },
+                    }),
+                    new Paragraph({
+                      children: [
+                        new TextRun({
+                          text:
+                            report.concerns ||
+                            "No concerns or struggles available.",
                           size: 14,
                         }),
                       ],
                       spacing: { after: 200 },
-                    }),
-                  ]),
-              new Paragraph({
-                children: [
-                  new TextRun({
-                    text: "Pray for",
-                    size: 16,
-                    bold: true,
-                  }),
-                ],
-                heading: "Heading3",
-                spacing: { after: 200 },
-              }),
-              ...(report.prayerRequests?.length
-                ? report.prayerRequests.map(
-                    (request) =>
-                      new Paragraph({
-                        children: [new TextRun({ text: request, size: 14 })],
-                        bullet: { level: 1 },
-                        spacing: { after: 200 },
-                      })
-                  )
-                : [
+                    })
+                  );
+                }
+
+                if (miscTitle === "Stories/Testimony") {
+                  sections.push(
                     new Paragraph({
                       children: [
                         new TextRun({
-                          text: "No prayer requests available.",
+                          text: "Stories / Testimony",
+                          size: 18,
+                          bold: true,
+                        }),
+                      ],
+                      heading: "Heading2",
+                      spacing: { after: 200 },
+                    }),
+                    new Paragraph({
+                      children: [
+                        new TextRun({
+                          text:
+                            report.story || "No story or testimony available.",
                           size: 14,
                         }),
                       ],
                       spacing: { after: 200 },
-                    }),
-                  ])
-            );
-          }
+                    })
+                  );
+                }
 
-          if (miscTitle === "Concerns/Struggles") {
-            sections.push(
-              new Paragraph({
-                children: [
-                  new TextRun({
-                    text: "Concerns / Struggles",
-                    size: 18,
-                    bold: true,
-                  }),
-                ],
-                heading: "Heading2",
-                spacing: { after: 200 },
+                return sections;
               }),
-              new Paragraph({
-                children: [
-                  new TextRun({
-                    text:
-                      report.concerns || "No concerns or struggles available.",
-                    size: 14,
-                  }),
-                ],
-                spacing: { after: 200 },
-              })
-            );
-          }
-
-          if (miscTitle === "Stories/Testimony") {
-            sections.push(
-              new Paragraph({
-                children: [
-                  new TextRun({
-                    text: "Stories / Testimony",
-                    size: 18,
-                    bold: true,
-                  }),
-                ],
-                heading: "Heading2",
-                spacing: { after: 200 },
-              }),
-              new Paragraph({
-                children: [
-                  new TextRun({
-                    text: report.story || "No story or testimony available.",
-                    size: 14,
-                  }),
-                ],
-                spacing: { after: 200 },
-              })
-            );
-          }
-          return sections;
-        }),
-      ]),
+            ]
+          ),
+        ]
+      ),
     ];
 
     return documentContent;
